Update to use UBI8, newer Strongswan, and not install Yum
Strongswan doesn't appear to be available in the UBI repositories and has some dependency issues in UBI 7 so I am proposing the switch to UBI 8 and adding the EPEL repo to install a newer Strongswan. This also removes some private keys installed from pygpgme package installed as a result of installing yum. I don't know if there are other dependency ramifications from upping Strongswan or switching to UBI 8 so this is merely a suggestion.